[svn] Concatenation with dates and booleans implemented and tested (with differences vs. Excel)

+++ src/classes/sej/internal/model/optimizer/consteval/EvalCell.java
@@ -51,7 +51,7 @@ public class EvalCell extends EvalShadow
 		if (null != constantValue) {
 			if (constantValue instanceof Boolean) {
 				boolean bool = ((Boolean) constantValue).booleanValue();
-				return type().adjustConstantValue( Double.valueOf( bool ? 1.0 : 0.0 ) );
+				return type().adjustConstantValue( Integer.valueOf( bool ? 1 : 0 ) );
 			}
 			return constantValue;
 		}
